Removing dependency on BARC PHP7-Nginx role
This caused issues in use-cases when this role was making changes to existing Monolith instances, such as generating server blocks for Monolith services.

Having this role as a dependency meant its tasks were always executed, which are quite time consuming.

Instead, this role will now a role requirement, in situations where a Monolith instance is being setup or configured.

To support this, templates and variables used from the Nginx role need to be contained in this role. Templates have been copied in, variables have either been given inline defaults (we can assume conventional ports, unlike BARC roles which are deliberately overly generic), hard-coded (index documents) or replaced with variables in this role.